# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tennis tournament",
        "description": "Manage participants and play the game",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/game/play": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Play a game",
                "operationId": "705e3fc0466fb5e5d04cdd2916df75b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "force": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game played successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The winner is John Doe with id = 123"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/game/reset": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Reset the game",
                "operationId": "0a3c7b9aa3b49239af3e407ca6fab261",
                "responses": {
                    "200": {
                        "description": "Game reseted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game reseted successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/game/replay": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Replay the game",
                "operationId": "f16ec84b6b7e3f0e92a1d04f0043fe27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "force": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game replayed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The winner is John Doe with id = 123"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/participants": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Get all participants",
                "operationId": "c164c364a97f0c8f58fa65374fea70d8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Participant"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Create a participant",
                "operationId": "d108be0bbe73bf5b055a969c7ed17032",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "skill": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 50
                                    },
                                    "strength": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 80
                                    },
                                    "speed": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 70
                                    },
                                    "reaction": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 90
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Participant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Participant created successfully"
                                        },
                                        "participant": {
                                            "$ref": "#/components/schemas/Participant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/participants/seed": {
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Seed participants",
                "operationId": "dac97f9b517332158579afcd2eda00dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 64
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Participants seeded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "64 participants created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/participants/clear": {
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Clear all participants",
                "operationId": "2cdfb1dcb0394ce8d17cd2415a93dcb5",
                "responses": {
                    "200": {
                        "description": "All participants deleted"
                    }
                }
            }
        },
        "/api/v1/tournaments": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Gets a list of tournaments",
                "description": "Returns a collection of tournaments, each with its associated participant. Can be filtered by gender and date.\n     *     **Filtering:**\n     *     * **gender:** Filters tournaments by participant's gender.\n     *     * **date:** Filters tournaments by creation date.",
                "operationId": "8f98aa3a558d7125411af976aa504612",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Participant's gender",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Tournament creation date",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tournaments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Tournament"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Participant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "skill": {
                        "type": "integer",
                        "format": "int32",
                        "example": 50
                    },
                    "strength": {
                        "type": "integer",
                        "format": "int32",
                        "example": 80
                    },
                    "speed": {
                        "type": "integer",
                        "format": "int32",
                        "example": 70
                    },
                    "reaction": {
                        "type": "integer",
                        "format": "int32",
                        "example": 90
                    },
                    "is_defeated": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-18T12:34:56+00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-18T12:34:56+00:00"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Tournament": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "gender": {
                        "type": "string",
                        "example": "female"
                    },
                    "participant_id": {
                        "type": "number",
                        "example": "256"
                    },
                    "participant": {
                        "$ref": "#/components/schemas/Participant"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-18T12:34:56+00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-18T12:34:56+00:00"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Game",
            "description": "Game"
        },
        {
            "name": "Participants",
            "description": "Participants"
        },
        {
            "name": "Tournaments",
            "description": "Tournaments"
        }
    ]
}